requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
hydra-core>=1.3.0
omegaconf>=2.3.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
tqdm>=4.65.0
wandb>=0.15.0